+++ app/src/main/java/io/github/droidkaigi/confsched2018/di/AppComponent.kt
@@ -12,6 +12,7 @@ import javax.inject.Singleton
     AndroidInjectionModule::class,
     AppModule::class,
     NetworkModule::class,
+    ApiModule::class,
     DatabaseModule::class,
     MainActivityBuilder::class,
     MapActivityBuilder::class,
