+++ app/src/release/java/io/github/droidkaigi/confsched2018/di/NetworkModule.kt
@@ -23,35 +23,4 @@ import javax.inject.Singleton
             .addNetworkInterceptor(HttpLoggingInterceptor()
                     .setLevel(HttpLoggingInterceptor.Level.BASIC))
             .build()
-
-    @Singleton @Provides @JvmStatic
-    fun provideRetrofit(okHttpClient: OkHttpClient): Retrofit {
-        return Retrofit.Builder()
-                .client(okHttpClient)
-                .baseUrl("https://droidkaigi.jp/2018/sessionize/")
-                .addConverterFactory(MoshiConverterFactory.create(Moshi.Builder()
-                        .add(ApplicationJsonAdapterFactory.INSTANCE)
-                        .add(LocalDateTime::class.java, LocalDateTimeAdapter())
-                        .build()))
-                .addCallAdapterFactory(RxJava2CallAdapterFactory.createAsync())
-                .build()
-    }
-
-    @Singleton @Provides @JvmStatic
-    fun provideDroidKaigiApi(retrofit: Retrofit): DroidKaigiApi {
-        return retrofit.create(DroidKaigiApi::class.java)
-    }
-
-    @Singleton @Provides @JvmStatic
-    fun provideFeedApi(): FeedApi = FeedFirestoreApi()
-
-    @Singleton @Provides @JvmStatic
-    fun provideSessionFeedbackApi(@RetrofitGoogleForm retrofit: Retrofit): SessionFeedbackApi {
-        return retrofit.create(SessionFeedbackApi::class.java)
-    }
-
-    @Singleton @Provides @JvmStatic
-    fun provideGithubApi(@RetrofitGithub retrofit: Retrofit): GithubApi {
-        return retrofit.create(GithubApi::class.java)
-    }
 }
